refactor: move app component to pages folder using atomic design
