+++ src/providers/react.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
 import '../index.css';
-import App from '../App';
+import App from '../pages/app';
 
 render(<App />, document.getElementById('root'));
